finish projcet

+++ Seagame_Ticket/app/Http/Controllers/StadiumController.php
@@ -16,19 +16,19 @@ class StadiumController extends Controller
         //
         // $data = Stadium::with('zone')->get();
         $data = Stadium::join('zones', 'stadia.zone_id', '=', 'zones.id')
-                ->select('stadia.*', 'zones.name as zone_name')
-                ->get();
-        return response()->json(['message'=>'This is a list of stadium','data'=>$data],200);
+            ->select('stadia.*', 'zones.name as zone_name')
+            ->get();
+        return response()->json(['message' => 'This is a list of stadium', 'data' => $data], 200);
     }
     public function store(Request $request)
     {
         //
-        $data= Stadium::create([
+        $data = Stadium::create([
             'name' => $request['name'],
             'zone_id' => $request['zone_id'],
 
         ]);
-        return response()->json(['message'=>'You data create successfully'],200);
+        return response()->json(['message' => 'You data create successfully'], 200);
     }
 
     /**
@@ -38,12 +38,15 @@ class StadiumController extends Controller
     {
         //
         $data = Stadium::find($id);
-        $data->update([
-            'name' => $request['name'],
-            'zone_id' => $request['zone_id'],
+        if ($data) {
+            $data->update([
+                'name' => $request['name'],
+                'zone_id' => $request['zone_id'],
 
-        ]);
-        return response()->json(['message'=>'You data update successfully'],200);
+            ]);
+            return response()->json(['message' => 'You data update successfully'], 200);
+        }
+        return  response()->json(['message' => "ID not fount"], 200);
     }
 
     /**
@@ -52,8 +55,12 @@ class StadiumController extends Controller
     public function destroy($id)
     {
         //
-        $data= Stadium::find($id);
-        $data->destroy($id);
-        return response()->json(['message'=>'You data deleted successfully'],200);
+        $data = Stadium::find($id);
+        if ($data) {
+            $data->destroy($id);
+            return response()->json(['message' => 'You data deleted successfully'], 200);
+        }
+        return response()->json(['message' => 'ID not found'], 200);
+
     }
 }
